Move WaitForService() into e2eservice
